localtest/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
pydantic==2.9.2
pyyaml>=6.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
